+++ packages/amplify-category-geo/src/provider-controllers/index.ts
@@ -6,6 +6,7 @@ import { addPlaceIndexResource, updatePlaceIndexResource, removePlaceIndexResour
 import { addMapResource, updateMapResource, removeMapResource } from './map';
 import { printer, prompter } from 'amplify-prompts';
 import { getServiceFriendlyName } from '../service-walkthroughs/resourceWalkthrough';
+import { TemplateMappings } from '../service-stacks/baseStack';
 
 /**
  * Entry point for creating a new Geo resource
@@ -121,3 +122,18 @@ const badServiceError = (service: string) => {
 export const insufficientInfoForUpdateError = (service: ServiceName) => {
   new Error(`Insufficient information to update ${getServiceFriendlyName(service)}. Please re-try and provide all inputs.`);
 }
+
+export const getTemplateMappings = async (context: $TSContext): Promise<TemplateMappings> => {
+  const Mappings: TemplateMappings = {
+    RegionMapping: {}
+  };
+  const providerPlugins = context.amplify.getProviderPlugins(context);
+  const providerPlugin = await import(providerPlugins[provider]);
+  const regionMapping = providerPlugin.getLocationRegionMapping();
+  Object.keys(regionMapping).forEach(region => {
+    Mappings.RegionMapping[region] = {
+      locationServiceRegion: regionMapping[region],
+    };
+  });
+  return Mappings;
+};