+++ packages/amplify-graphql-auth-transformer/src/__tests__/group-auth.test.ts
@@ -1,0 +1,97 @@
+import { AuthTransformer } from '@aws-amplify/graphql-auth-transformer';
+import { ModelTransformer } from '@aws-amplify/graphql-model-transformer';
+import { GraphQLTransform } from '@aws-amplify/graphql-transformer-core';
+import { ResourceConstants } from 'graphql-transformer-common';
+import { AppSyncAuthConfiguration } from '@aws-amplify/graphql-transformer-interfaces';
+
+test('happy case with static groups', () => {
+  const authConfig: AppSyncAuthConfiguration = {
+    defaultAuthentication: {
+      authenticationType: 'AMAZON_COGNITO_USER_POOLS',
+    },
+    additionalAuthenticationProviders: [],
+  };
+  const validSchema = `
+  type Post @model @auth(rules: [{allow: groups, groups: ["Admin", "Dev"]}]) {
+    id: ID!
+    title: String!
+    createdAt: String
+    updatedAt: String
+  }`;
+  const transformer = new GraphQLTransform({
+    authConfig,
+    transformers: [
+      new ModelTransformer(),
+      new AuthTransformer({
+        authConfig,
+        addAwsIamAuthInOutputSchema: false,
+      }),
+    ],
+  });
+  const out = transformer.transform(validSchema);
+  expect(out).toBeDefined();
+  expect(out.rootStack!.Resources![ResourceConstants.RESOURCES.GraphQLAPILogicalID].Properties.AuthenticationType).toEqual(
+    'AMAZON_COGNITO_USER_POOLS',
+  );
+});
+
+test('happy case with dynamic groups', () => {
+  const authConfig: AppSyncAuthConfiguration = {
+    defaultAuthentication: {
+      authenticationType: 'AMAZON_COGNITO_USER_POOLS',
+    },
+    additionalAuthenticationProviders: [],
+  };
+  const validSchema = `
+    type Post @model @auth(rules: [{allow: groups, groupsField: "groups"}]) {
+        id: ID!
+        title: String!
+        groups: [String]
+        createdAt: String
+        updatedAt: String
+    }
+    `;
+  const transformer = new GraphQLTransform({
+    authConfig,
+    transformers: [
+      new ModelTransformer(),
+      new AuthTransformer({
+        authConfig,
+        addAwsIamAuthInOutputSchema: false,
+      }),
+    ],
+  });
+  const out = transformer.transform(validSchema);
+  expect(out).toBeDefined();
+  expect(out.rootStack!.Resources![ResourceConstants.RESOURCES.GraphQLAPILogicalID].Properties.AuthenticationType).toEqual(
+    'AMAZON_COGNITO_USER_POOLS',
+  );
+});
+
+test('validation on @auth on a non-@model type', () => {
+  const authConfig: AppSyncAuthConfiguration = {
+    defaultAuthentication: {
+      authenticationType: 'AMAZON_COGNITO_USER_POOLS',
+    },
+    additionalAuthenticationProviders: [],
+  };
+  const invalidSchema = `
+    type Post @auth(rules: [{allow: groups, groupsField: "groups"}]) {
+        id: ID!
+        title: String!
+        group: String
+        createdAt: String
+        updatedAt: String
+    }`;
+  const transformer = new GraphQLTransform({
+    authConfig,
+    transformers: [
+      new ModelTransformer(),
+      new AuthTransformer({
+        authConfig,
+        addAwsIamAuthInOutputSchema: false,
+      }),
+    ],
+  });
+  expect(() => transformer.transform(invalidSchema)).toThrowError('Types annotated with @auth must also be annotated with @model.');
+});