+++ packages/amplify-graphql-auth-transformer/src/index.ts
@@ -1,0 +1,4 @@
+export * from './graphql-auth-transformer';
+export * from './utils/constants';
+export * from './utils/definitions';
+export { AccessControlMatrix } from './accesscontrol';