+++ packages/amplify-category-auth/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.38.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-auth@2.38.3...amplify-category-auth@2.38.4) (2021-10-10)
+
+**Note:** Version bump only for package amplify-category-auth
+
+
+
+
+
+## [2.38.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-auth@2.38.2...amplify-category-auth@2.38.3) (2021-10-06)
+
+**Note:** Version bump only for package amplify-category-auth
+
+
+
+
+
 ## [2.38.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-auth@2.38.1...amplify-category-auth@2.38.2) (2021-09-27)